+++ components/simulation-page/simulation-menu/index.ts
@@ -1,0 +1,1 @@
+export { SimulationMenuBar } from "./SimulationMenuBar";
