+++ components/simulation-page/simulation-menu/legende/index.ts
