+++ package.json
@@ -70,9 +70,9 @@
     "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
-    "@babel/node": "^7.8.7",
     "@babel/cli": "^7.8.4",
     "@babel/core": "^7.9.0",
+    "@babel/node": "^7.8.7",
     "@babel/preset-env": "^7.9.5",
     "@types/node": "^13.13.2",
     "@types/react": "^16.9.34",
